Other fixes necessary to compile with a C++ compiler: * Subtlety different semantics for extern keyword. * Inability to do a forward declaration of a static variable. * Structs must be initialized in the same order as declared. * Cannot initialize a fixed size character array with a quoted string. * Cannot link with a library by stating the name of the .a file. * Struct initialization sometimes needs explicit cast.

+++ ep/ep_stat.c
@@ -66,7 +66,7 @@ ep_stat_from_errno(int uerrno)
 	if (uerrno == 0)
 		return EP_STAT_OK;
 	return EP_STAT_NEW(ERROR, EP_REGISTRY_EPLIB,
-			EP_STAT_MOD_ERRNO, (unsigned long) uerrno);
+			EP_STAT_MOD_ERRNO, (uint32_t) uerrno);
 }
